# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Paseban Kawis",
        "description": "Dokumentasi API Paseban Kawis",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/modul": {
            "get": {
                "tags": [
                    "Modul"
                ],
                "summary": "List semua modul dengan data kategori",
                "operationId": "e3e1df3d08b646b167330a00745552d6",
                "responses": {
                    "200": {
                        "description": "List modul dengan data kategori",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Modul"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Modul"
                ],
                "summary": "Tambah modul baru (admin only, role_id=1)",
                "operationId": "85b68db46dded71e6e304d4ef04af9cd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "judul_modul",
                                    "category_modul_id",
                                    "link_video",
                                    "deskripsi_modul"
                                ],
                                "properties": {
                                    "judul_modul": {
                                        "type": "string",
                                        "example": "Dasar Desain Produk"
                                    },
                                    "category_modul_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "link_video": {
                                        "type": "string",
                                        "example": "https://youtu.be/xxxx"
                                    },
                                    "pdf": {
                                        "type": "file",
                                        "format": "binary"
                                    },
                                    "foto": {
                                        "type": "file",
                                        "format": "binary"
                                    },
                                    "deskripsi_modul": {
                                        "type": "string",
                                        "example": "Penjelasan modul..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Modul berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Modul"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/modul/{id}": {
            "get": {
                "tags": [
                    "Modul"
                ],
                "summary": "Detail modul dengan data kategori",
                "operationId": "34bd3053d3bee7853841f025205e81d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail modul dengan data kategori",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Modul"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Modul"
                ],
                "summary": "Update modul (admin only, role_id=1)",
                "operationId": "3f3a8067656ae5b1689b16cc857faa12",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "judul_modul": {
                                        "type": "string",
                                        "example": "Dasar Desain Produk"
                                    },
                                    "category_modul_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "link_video": {
                                        "type": "string",
                                        "example": "https://youtu.be/xxxx"
                                    },
                                    "pdf": {
                                        "type": "file",
                                        "format": "binary"
                                    },
                                    "foto": {
                                        "type": "file",
                                        "format": "binary"
                                    },
                                    "deskripsi_modul": {
                                        "type": "string",
                                        "example": "Penjelasan modul..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Modul berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Modul"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Modul"
                ],
                "summary": "Hapus modul (admin only, role_id=1)",
                "operationId": "53f0e548139a564d541a984b75b87af4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Modul berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Modul deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CategoryModul": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nama": {
                        "type": "string",
                        "example": "Design Product"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Modul": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "judul_modul": {
                        "type": "string",
                        "example": "Dasar Desain Produk"
                    },
                    "category_modul_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "link_video": {
                        "type": "string",
                        "example": "https://youtu.be/xxxx"
                    },
                    "path_pdf": {
                        "type": "string",
                        "example": "storage/modul_pdfs/abc.pdf"
                    },
                    "foto": {
                        "type": "string",
                        "example": "storage/modul_fotos/abc.jpg"
                    },
                    "deskripsi_modul": {
                        "type": "string",
                        "example": "Penjelasan modul..."
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "category_modul": {
                        "$ref": "#/components/schemas/CategoryModul"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}